requirements.txt
tensorflow==2.12.0
numpy
matplotlib
scikit-learn
opencv-python
pillow
